#288: renames submodule from jfxbfr to clv, introduces new clv-app module, adds some performance optimisations
