+++ app/src/main/scala/app/logorrr/model/LogEntrySelector.scala
@@ -1,6 +1,6 @@
 package app.logorrr.model
 
-import app.logorrr.jfxbfr.ElementSelector
+import app.logorrr.clv.ElementSelector
 import javafx.beans.property.SimpleIntegerProperty
 
 case class LogEntrySelector(selectedLineNumberProperty: SimpleIntegerProperty) extends ElementSelector[LogEntry] {
