+++ app/src/main/scala/app/logorrr/views/MutFilter.scala
@@ -1,6 +1,6 @@
 package app.logorrr.views
 
-import app.logorrr.jfxbfr.color.ColorMatcher
+import app.logorrr.clv.color.ColorMatcher
 import app.logorrr.views.search.predicates.DescriptivePredicate
 import javafx.beans.property.{BooleanProperty, SimpleBooleanProperty, SimpleObjectProperty}
 import javafx.scene.paint.Color
