+++ clv/src/main/scala/app/logorrr/clv/ElementSelector.scala
@@ -1,0 +1,9 @@
+package app.logorrr.clv
+
+/**
+ * Implement this function to define what action is performed if an element is selected
+ * @tparam A
+ */
+trait ElementSelector[A] {
+  def select(a: A): Unit
+}
