+++ clv/src/main/scala/app/logorrr/clv/color/ColorChozzer.scala
@@ -1,0 +1,9 @@
+package app.logorrr.clv.color
+
+import javafx.scene.paint.Color
+
+trait ColorChozzer[A] {
+
+  def calc(a: A): Color
+
+}
