+++ clv/src/main/scala/app/logorrr/clv/color/ColorMatcher.scala
@@ -1,0 +1,10 @@
+package app.logorrr.clv.color
+
+import javafx.scene.paint.Color
+
+trait ColorMatcher {
+
+  def matches(searchTerm: String): Boolean
+
+  def getColor: Color
+}
