+++ clv/src/test/scala/app/logorrr/clv/ColorUtilSpec.scala
@@ -1,0 +1,14 @@
+package app.logorrr.clv
+
+import javafx.scene.paint.Color
+import org.scalatest.wordspec.AnyWordSpec
+
+class ColorUtilSpec extends AnyWordSpec {
+
+  "Convert color to string and back" in {
+    val color = Color.WHITE
+    val c1 = Color.web(color.toString)
+    assert(color == c1)
+  }
+
+}
