--- jfxbfr/src/main/scala/app/logorrr/jfxbfr/color/ColorChozzer.scala
@@ -1,9 +1,0 @@
-package app.logorrr.jfxbfr.color
-
-import javafx.scene.paint.Color
-
-trait ColorChozzer[A] {
-
-  def calc(a: A): Color
-
-}
