--- jfxbfr/src/main/scala/app/logorrr/jfxbfr/color/ColorMatcher.scala
@@ -1,10 +1,0 @@
-package app.logorrr.jfxbfr.color
-
-import javafx.scene.paint.Color
-
-trait ColorMatcher {
-
-  def matches(searchTerm: String): Boolean
-
-  def getColor: Color
-}
